feat: Terminado

+++ src/main/java/io/gitHub/CristianoAlberto/services/GuestService.java
@@ -2,12 +2,13 @@ package io.gitHub.CristianoAlberto.services;
 
 import io.gitHub.CristianoAlberto.models.GuestEntity;
 import io.gitHub.CristianoAlberto.repositories.GuestRepository;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.concurrent.CompletableFuture;
 
 @Service
@@ -26,20 +27,40 @@ public class GuestService {
 
     @Async
     public CompletableFuture<GuestEntity> getById(Integer id) {
-        return CompletableFuture.completedFuture(guestRepository.getReferenceById(id));
+        Optional<GuestEntity> guest = guestRepository.findById(id);
+        if(guest.isPresent()){
+            return CompletableFuture.completedFuture(guest.get());
+        }else {
+            return CompletableFuture.failedFuture(new EntityNotFoundException());
+        }
     }
-
     @Async
     public CompletableFuture<GuestEntity> createGuest(GuestEntity guest) {
         return CompletableFuture.completedFuture(guestRepository.save(guest));
     }
-
+    @Async
+    public CompletableFuture<GuestEntity> updateGuest(GuestEntity guest,Integer id){
+        return CompletableFuture.supplyAsync(() -> {
+            Optional<GuestEntity> guestOptional = guestRepository.findById(id);
+            if (guestOptional.isPresent()) {
+                GuestEntity existingGuest = guestOptional.get();
+                existingGuest.setName(guest.getName());
+                existingGuest.setNumber(guest.getNumber());
+                existingGuest.setGender(guest.getGender());
+                existingGuest.setConfirmation(guest.getConfirmation());
+                return guestRepository.save(existingGuest);
+            } else {
+                throw new EntityNotFoundException();
+            }
+        });
+    }
     @Async
     public CompletableFuture<Boolean> deleteGuest(Integer id) {
-        try {
+        Optional<GuestEntity> guestOptional = guestRepository.findById(id);
+        if (guestOptional.isPresent()) {
             guestRepository.deleteById(id);
             return CompletableFuture.completedFuture(true);
-        } catch (EmptyResultDataAccessException e) {
+        } else {
             return CompletableFuture.completedFuture(false);
         }
     }
